requirements.txt
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
exceptiongroup==1.3.0
greenlet==3.2.4
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-groq==0.3.7
langchain-openai==0.3.32
langchain-text-splitters==0.3.10
langsmith==0.4.21
openai==1.102.0
orjson==3.11.3
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
zstandard==0.24.0
